Add files via upload
Run from the APIC cli. This tool is used to run leaf/spine cli commands remotely from the apic. For example, if a CLI command needed to be run on 100 leafs this tool would do this. Its most useful for output that can't be pulled from the object model but rather requires a show command or some linux command.